% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqminer.R
\name{createSingleChromosomeVCFIndex}
\alias{createSingleChromosomeVCFIndex}
\title{Create a single chromosome index}
\usage{
createSingleChromosomeVCFIndex(fileName, indexFileName = NULL)
}
\arguments{
\item{fileName}{character, represents an input VCF file (Bgzipped, with Tabix index)}

\item{indexFileName}{character, by default, create `fileName`.scIdx}
}
\value{
indexFileName if success, or NULL is failed
}
\description{
Create a single chromosome index
}
\examples{
fileName = system.file("vcf/all.anno.filtered.extract.vcf.gz", package = "seqminer")
cfh <- createSingleChromosomeVCFIndex(fileName)
}
\seealso{
http://zhanxw.com/seqminer/ for online manual and examples
}
