% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqminer.R
\name{download.annotation.resource}
\alias{download.annotation.resource}
\title{Download annotation resources to a directory}
\usage{
download.annotation.resource(outputDirectory)
}
\arguments{
\item{outputDirectory}{the directory to store annotation resources}
}
\value{
will not return anything
}
\description{
Download annotation resources to a directory
}
\examples{
\dontrun{
download.annotation.resource("/tmp")
}
}
