% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqminer.R
\name{isBGENSupported}
\alias{isBGENSupported}
\title{Check if required BGEN library is installed}
\usage{
isBGENSupported()
}
\value{
bool
}
\description{
Check if required BGEN library is installed
}
\examples{
isBGENSupported()
}
