% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqminer.R
\name{isDirWritable}
\alias{isDirWritable}
\title{Test whether directory is writable}
\usage{
isDirWritable(outDir)
}
\arguments{
\item{outDir}{the name of the directory}
}
\value{
TRUE if the file is writable
isDirWritable("~")
}
\description{
Test whether directory is writable
}

