% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqminer.R
\name{isTabixRange}
\alias{isTabixRange}
\title{Check if the inputs are valid tabix range such as chr1:2-300}
\usage{
isTabixRange(range)
}
\arguments{
\item{range}{characer vector}
}
\description{
Check if the inputs are valid tabix range such as chr1:2-300
}
\examples{
valid <- isTabixRange(c("chr1:1-200", "X:1", "1:100-100", "chr1", "1:1-20,1:30-40"))
stopifnot(all(valid))
invalid <- isTabixRange(c(":1", "chr1::", ":-"))
stopifnot(all(!invalid))
}

