% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqminer.R
\name{isURL}
\alias{isURL}
\title{Check if the input is url e.g. http:// or ftp://}
\usage{
isURL(fileName)
}
\arguments{
\item{fileName}{character vector}
}
\description{
Check if the input is url e.g. http:// or ftp://
}
\keyword{internal}
