% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqminer.R
\name{makeAnnotationParameter}
\alias{makeAnnotationParameter}
\title{Construct a usable set of annotation parameters}
\usage{
makeAnnotationParameter(param = NULL)
}
\arguments{
\item{param}{a list of annotation elements}
}
\value{
list, a complete list of supported parameters
}
\description{
Construct a usable set of annotation parameters
}

