% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqminer.R
\name{makeSureBGENSupported}
\alias{makeSureBGENSupported}
\title{Stop if required BGEN library is not installed}
\usage{
makeSureBGENSupported()
}
\value{
bool
}
\description{
Stop if required BGEN library is not installed
}
\examples{
makeSureBGENSupported()
}
