% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMakefile.R
\name{newJob}
\alias{newJob}
\title{Create a new job}
\usage{
newJob(id, cmd, outFile, depend = NULL)
}
\arguments{
\item{id}{character, job ids.}

\item{cmd}{character, commands to run}

\item{outFile}{character, the output file names after command are run successfully}

\item{depend}{character vector, specify the prerequisite files (e.g. outFile from other jobs)}
}
\description{
Create a new job
}
\examples{
j1 <- newJob('id1', 'cmd out1', 'out1')
j2 <- newJob('id2', 'cmd out2', 'out2', depend = 'out1')
}
