% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genMakefile.R
\name{newWorkflow}
\alias{newWorkflow}
\title{Create a new workflow}
\usage{
newWorkflow(name)
}
\arguments{
\item{name}{character, specify the name of the workflow}
}
\description{
Create a new workflow
}
\examples{
w <- newWorkflow("wf")
}

