% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqminer.R
\name{openPlink}
\alias{openPlink}
\title{Open binary PLINK files}
\usage{
openPlink(fileName)
}
\arguments{
\item{fileName}{character, represents the prefix of PLINK input file}
}
\value{
an PLINK file object with class name ("PlinkFile")
}
\description{
Open binary PLINK files
}
\examples{
fileName = system.file("plink/all.anno.filtered.extract.bed", package = "seqminer")
fileName = sub(fileName, pattern = ".bed", replacement = "")
plinkObj <- openPlink(fileName)
str(plinkObj)
}
