% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqminer.R
\docType{package}
\name{SEQMINER}
\alias{seqminer}
\alias{seqminer-package}
\title{Efficiently Read Sequencing Data (VCF format, METAL format) into R}
\description{
SeqMiner provides functions to easily load Variant Call Format (VCF) or METAL format into R
}
\details{
The aim of this package is to save your time parsing large text file.
That means data processing time can be saved for other researches.
This packages requires Bgzip compressed and Tabix indexed files as input.
If input files contaians annotation by TabAnno (),
it is possible to extract information at the unit of genes.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://zhanxw.github.io/seqminer/}
  \item Report bugs at \url{https://github.com/zhanxw/seqminer/issues}
}

}
\author{
\strong{Maintainer}: Xiaowei Zhan \email{zhanxw@gmail.com}

Authors:
\itemize{
  \item Dajiang Liu \email{dajiang.liu@gmail.com}
}

Other contributors:
\itemize{
  \item Attractive Chaos \email{attractor@live.co.uk} (We have used the following software and made minimal necessary changes: Tabix, Heng Li <lh3@live.co.uk> (MIT license). We removed standard IO related functions, e.g. printf, fprintf ; also changed its un-safe pointer arithmetics.) [copyright holder]
  \item Broad Institute / Massachusetts Institute of Technology [copyright holder]
  \item Genome Research Ltd (GRL) [copyright holder]
  \item Facebook, Inc [copyright holder]
  \item D. Richard Hipp [copyright holder]
}

}
