\name{validateAnnotationParameter}
\alias{validateAnnotationParameter}
\title{Validate annotate parameter is valid}
\usage{
validateAnnotationParameter(param, debug = FALSE)
}
\arguments{
  \item{param}{a list of annotation elements}

  \item{debug}{show extra debug information or not}
}
\value{
list, first element is TRUE/FALSE if parameter is
valid/invalid;
}
\description{
Validate annotate parameter is valid
}

