% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqminer.R
\name{verifyFilename}
\alias{verifyFilename}
\title{validate the inVcf can be created, and outVcf can be write to.
will stop if any error occurs}
\usage{
verifyFilename(inVcf, outVcf)
}
\arguments{
\item{inVcf}{input file}

\item{outVcf}{output file}
}
\description{
validate the inVcf can be created, and outVcf can be write to.
will stop if any error occurs
}
