\name{hasIndex}
\alias{hasIndex}
\title{Check input file has tabix index}
\usage{
hasIndex(fileName)
}
\arguments{
  \item{fileName}{character vector, representing file names
  an input file name}
}
\value{
TRUE if an index file with ".tbi" or "bci" exists, and
created later than VCF/BCF file
}
\description{
Check input file has tabix index
}
\keyword{internal}

