\name{makeAnnotationParameter}
\alias{makeAnnotationParameter}
\title{Construct a usable set of annotation parameters}
\usage{
makeAnnotationParameter(param = NULL)
}
\arguments{
  \item{param}{a list of annotation elements}
}
\value{
list, a complete list of supported parameters
}
\description{
Construct a usable set of annotation parameters
}

