% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqminer.R
\docType{package}
\name{SeqMiner}
\alias{SeqMiner}
\alias{SeqMiner-package}
\title{Efficiently Read Sequencing Data (VCF format, METAL format) into R}
\description{
SeqMiner provides functions to easily load Variant Call Format (VCF) or METAL format into R
}
\details{
The aim of this package is to save your time parsing large text file.
That means data processing time can be saved for other researches.
This packages requires Bgzip compressed and Tabix indexed files as input.
If input files contaians annotation by TabAnno (),
it is possible to extract information at the unit of genes.
}

